refactor: extract employee list base activity
